feat(VsToast): move toast-plugin types to seperate types file
