+++ packages/vlossom/src/components/vs-toast/VsToast.vue
@@ -20,11 +20,11 @@
 import { PropType, computed, defineComponent, ref, toRef, toRefs } from 'vue';
 import { UIState } from '@/declaration';
 import { useColorScheme } from '@/composables';
-import { ToastInfo } from '@/declaration';
 import { VsIcon } from '@/icons';
 import { store } from '@/store';
 
 import type { VsComponent } from '@/declaration';
+import type { ToastInfo } from '@/plugins';
 
 const name = 'VsToast';
 export default defineComponent({
