+++ packages/vlossom/src/declaration/types.ts
@@ -153,30 +153,3 @@ export type CssPosition = (typeof CSS_POSITION)[number];
 export type Placement = (typeof PLACEMENTS)[number];
 
 export type Size = (typeof SIZES)[number];
-
-export interface ToastInfo {
-    id: string;
-    text: string;
-    autoClose: boolean;
-    duration?: number;
-    placement: Exclude<Placement, 'left' | 'right'>;
-    align: Align;
-    colorScheme?: ColorScheme;
-    state?: Exclude<UIState, UIState.Idle | UIState.Selected>;
-}
-
-export type ToastOptions = {
-    autoClose?: boolean;
-    timeout?: number;
-    placement?: Exclude<Placement, 'left' | 'right'>;
-    align?: Align;
-    colorScheme?: ColorScheme;
-};
-
-export interface ToastPlugin {
-    show(text: string, toastOptions?: ToastOptions): void;
-    success(text: string, toastOptions?: ToastOptions): void;
-    info(text: string, toastOptions?: ToastOptions): void;
-    error(text: string, toastOptions?: ToastOptions): void;
-    warn(text: string, toastOptions?: ToastOptions): void;
-}
