+++ packages/vlossom/src/plugins/index.ts
@@ -1,1 +1,2 @@
-export { toastPlugin } from './toast-plugin';
+export * from './toast-plugin/types';
+export * from './toast-plugin';
